==> 8th SEM/FYP - Done/1 Skill_Extracting_for_All_Resumes.ipynb <==
!pip install -U spacy
!pip install jsonlines

!python -m spacy download en_core_web_lg

# **Imports**

import spacy
from spacy.pipeline import EntityRuler
from spacy.lang.en import English
from spacy.tokens import Doc

#gensim
import gensim
from gensim import corpora

#Data loading/ Data manipulation
import pandas as pd
import numpy as np
import jsonlines

#nltk
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
nltk.download(['stopwords','wordnet'])

#warning
import warnings 
warnings.filterwarnings('ignore')

## **Loading Spacy Model**

nlp = spacy.load("en_core_web_lg")
skill_pattern_path = "./jz_skill_patterns.jsonl"

# **Entity Ruler**


ruler = nlp.add_pipe("entity_ruler")
ruler.from_disk(skill_pattern_path)
nlp.pipe_names

def unique_skills(x):
    return list(set(x))

def get_skills(text):
    doc = nlp(text)
    myset = []
    subset = []
    for ent in doc.ents:
        if ent.label_ == "SKILL":
            subset.append(ent.text)
    myset.append(subset)
    return unique_skills(subset)

data = pd.read_csv("Resume.csv")
data

def clean_resumes(data):
    review = re.sub(
        '(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)|^rt|http.+?"',
        " ",
        data,
    )
    review = review.lower()
    review = review.split()
    lm = WordNetLemmatizer()
    review = [
        lm.lemmatize(word)
        for word in review
        if not word in set(stopwords.words("english"))
    ]
    review = " ".join(review)
    return review

#cleaning data
data['cleaned_resume'] = data.resume.apply(lambda x: clean_resumes(x))

data['skills'] = data.cleaned_resume.apply(lambda x: get_skills(x))
data

data.to_csv('./resume_with_skills.csv')
