==> 8th SEM/FYP - Done/7 For_JD_recommend_resume.ipynb <==
!pip install -U gensim

import pandas as pd
import numpy as np
import gensim
from sklearn.linear_model import LogisticRegression
from gensim.models.doc2vec import TaggedDocument
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import roc_curve, roc_auc_score
from sklearn.naive_bayes import MultinomialNB
from sklearn.preprocessing import MinMaxScaler
from sklearn.pipeline import Pipeline
from sklearn.multiclass import OneVsRestClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn import metrics
from sklearn.svm import SVC
from tqdm import tqdm
tqdm.pandas(desc="progress-bar")
from gensim.models import Doc2Vec
import multiprocessing
from sklearn import utils
from sklearn.model_selection import train_test_split
from gensim.parsing.preprocessing import remove_stopwords
from gensim.utils import simple_preprocess
from gensim.parsing.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer
import re
import nltk

import itertools
from gensim.similarities import MatrixSimilarity
from gensim.parsing.preprocessing import preprocess_documents
from gensim.parsing.preprocessing import preprocess_string

nltk.download('wordnet')

resumeDf = pd.read_csv('resume_with_skills.csv')
resumeDf.drop(resumeDf.columns[[0]], axis = 1, inplace = True)
resumeDf #access row by id - 1

jdDf = pd.read_csv('jd_with_skills.csv')
jdDf.drop(jdDf.columns[[0]], axis = 1, inplace = True)
jdDf

jdsimDf = pd.read_csv('jd_similar.csv')
jdsimDf

tagged_corpus = []
for i in range(len(resumeDf)):
  tagged_corpus.append(TaggedDocument(resumeDf['cleaned_resume'][i].split(' '), [resumeDf['id'][i]]))

model = Doc2Vec(tagged_corpus,
                dm=0, 
                vector_size=500, 
                window=5, 
                min_count=1, 
                epochs=10, 
                hs=1)

model.save('./d2vResume.model')

from scipy import spatial
import ast
import math 

def get_cosine(a, b):
  result = 1 - spatial.distance.cosine(a, b)
  return result

def get_sim(given, required):
  score = 0
  given = ast.literal_eval(given)
  required = ast.literal_eval(required)
  given = set(given)
  for x in required:
    if x in given:
      score = score + 1
  result = 0
  if len(required) != 0:
    result = ((score*1.0)/len(required))
  return result

def get_count(cat, df, resumes):

  c = 0
  d = 0
  i = 0

  for index, row in df.iterrows():
    id = row[0]
    i += 1
    if cat == resumes['category'][id-1]:
      d += 1
      if(i==10):
        c = d

  return [c, d]

list1 = [1, 3, 5, 7, 9]
list2 = [1, 3, 5, 7, 9]

result = pd.DataFrame()
result['id'] = ''
result['cos_10'] = ''
result['cos_20'] = ''
result['sim_10'] = ''
result['sim_20'] = ''

for i in list1:
  for j in list2:
    result['com_' + str(i) + '_' + str(j) + '_10' ] = ''
    result['com_' + str(i) + '_' + str(j) + '_20' ] = ''

result

tempDf = pd.DataFrame()
tempDf['id'] = ""
tempDf['cosine_sim'] = ""
tempDf['skills_sim'] = ""

for i in list1:
  for j in list2:
    tempDf['com_' + str(i) + '_' + str(j)] = ''

tempDf

for k in range(len(jdDf)):
  jd1 = jdDf['cleaned_jd'][k].split(' ')
  jd1 = model.infer_vector(jd1)
  jd1_skills = jdDf['skills'][k]
  cat = jdDf['category'][k]
  idd = jdDf['id'][k]

  tempDf = pd.DataFrame()
  tempDf['id'] = ""
  tempDf['cosine_sim'] = ""
  tempDf['skills_sim'] = ""

  for i in list1:
    for j in list2:
      tempDf['com_' + str(i) + '_' + str(j)] = ''

  for i in range(len(resumeDf)):
    id = resumeDf['id'][i]
    resume_vector = model.infer_vector(resumeDf['cleaned_resume'][i].split(' '))
    cosine = get_cosine(jd1, resume_vector)
    skills = get_sim(resumeDf['skills'][i] , jdsimDf['merged_skills'][j])

    # print(cosine, skills)

    listt = [id, cosine, skills]

    for i in list1:
      for j in list2:
        listt.append((i*cosine + j*skills)/(i+j))
    
    tempDf.loc[len(tempDf.index)] = listt

  cc = get_count(cat, tempDf.sort_values(by = ['cosine_sim'], ascending = False).head(20), resumeDf)
  cc_10 = cc[0]
  cc_20 = cc[1]

  dd = get_count(cat, tempDf.sort_values(by = ['skills_sim'], ascending = False).head(20), resumeDf)
  dd_10 = dd[0]
  dd_20 = dd[1]

  listtt = [idd, cc_10, cc_20, dd_10, dd_20]

  for i in list1:
    for j in list2:
      pp = get_count(cat, tempDf.sort_values(by=['com_' + str(i) + '_' + str(j)], ascending = False).head(20), resumeDf)
      listtt.append(pp[0])
      listtt.append(pp[1])

  if k%10 == 0:
    print(k)

  result.loc[len(result.index)] = listtt
result

get_count(resumeDf['category'][0], tempDf.sort_values(by = ['skills_sim'], ascending = False).head(20), jdDf)


for row in t2.head(20).iterrows():
  print(jdDf['category'][row[0]])

resumeDf['category'][200]

result.to_csv('jd_recommend.csv', index = False)
